refactor: decompose 652-line get-started into focused modules
Split app/(app)/dashboard/(components)/get-started.tsx into a get-started/
folder (import path unchanged via index.tsx):

- steps.ts: pure step definitions + completion rules, now unit-tested
- use-onboarding.ts: queries, onboarding mutation, derived step state and
  the auto-complete / legacy-account effects (query keys unchanged, still
  shared with verify-email-alert and account-deletion-alert)
- step-actions.tsx + plan-picker.tsx: per-step CTAs and the Free/Pro chooser
- register-help-dialog.tsx, skeleton.tsx: presentational pieces
- index.tsx: the timeline card view (~180 lines)
- bg-gradient-to-br -> bg-linear-to-br (Tailwind v4 canonical name)

Build, 25 unit tests (6 new), and 5 e2e green.
